requirements.txt
streamlit
groq
python-dotenv
requests
SpeechRecognition
sounddevice
soundfile
newsapi-python